% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_UAuPA_to_confmat.R
\name{add_UAuPA_to_confmat}
\alias{add_UAuPA_to_confmat}
\title{Add User's and Producer's Accuracy to a confusion matrix.}
\usage{
add_UAuPA_to_confmat(cm, as.int = F)
}
\arguments{
\item{cm}{A confusion matrix (table).}

\item{as.int}{If TRUE, accuracies are multiplied with 100 
and converted to integers.}
}
\value{
A table with the User's and Producer's Accuracies 
added as last column and row.
}
\description{
...
}
\examples{
\dontrun{
cm <- rbind(c(8, 2, 4),
            c(1, 7, 0),
            c(1, 1, 6))
rownames(cm) <- colnames(cm) <- c("a", "b", "c")
cm <- as.table(cm)
add_UAuPA_to_confmat(cm, TRUE)
}
}

