% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mxMisc.R
\name{addPaletteFun}
\alias{addPaletteFun}
\title{Add palette to map-x style list}
\usage{
addPaletteFun(sty, pal)
}
\arguments{
\item{sty}{map-x style}

\item{pal}{name of palette to use}
}
\description{
Update a style list with a palette, using the defined scale type : continuous or discrete.
}

