% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mxVtDep.R
\name{addVectorTiles}
\alias{addVectorTiles}
\title{Add vector tiles for a given PGRestAPI postgres endpoint.}
\usage{
addVectorTiles(map, protocol = "http", url = "localhost", port = 3030,
  table = NULL, dataColumns = NULL, geomColumn = "geom",
  idColumn = "gid", id = NULL, group = "default", debug = FALSE,
  zIndex = 100, onLoadFeedback = c("once", "never", "always"))
}
\arguments{
\item{map}{Leaflet map object}

\item{urlTemplate}{Url template for a given PGRestAPI endpoint.}
}
\description{
Add vector tiles for a given PGRestAPI postgres endpoint.
}

