% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_functions.R
\name{box}
\alias{box}
\title{Return the first N rows and columns of a data.frame or matrix}
\usage{
box(data, rows = 6, cols = rows)

box(data, rows = 6, cols = rows)
}
\arguments{
\item{data}{A data.frame, matrix or data type that has both rows and columns}

\item{rows}{The number of rows to be returned, similar to head(). Default value is 6.}

\item{cols}{The number of columns to be returned. Defaults to the same value as rows to maintain a square format.}

\item{data}{A data.frame, matrix or data type that has both rows and columns}

\item{rows}{The number of rows to be returned, similar to head(). Default value is 6.}

\item{cols}{The number of columns to be returned. Defaults to the same value as rows to maintain a square format.}
}
\value{
An object (usually) like data but generally smaller.

An object (usually) like data but generally smaller.
}
\description{
Return the first N rows and columns of a data.frame or matrix

Return the first N rows and columns of a data.frame or matrix
}
\details{
A two dimensional version of head() that is useful for working with wide datasets. Unlike head() and tail() this is not a generic function and cannot be extended to other classes.

A two dimensional version of head() that is useful for working with wide datasets. Unlike head() and tail() this is not a generic function and cannot be extended to other classes.
}
\examples{
box(iris)
box(mtcars, rows = 10, cols = 3)

box(matrix(0, nrow=3, ncol=3), 2)
box(iris)
box(mtcars, rows = 10, cols = 3)

box(matrix(0, nrow=3, ncol=3), 2)
}
