% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{calculate_net}
\alias{calculate_net}
\title{Calculates net migration from flow data.table}
\usage{
calculate_net(
  dt,
  col,
  by,
  o = "fromdist",
  d = "todist",
  long = FALSE,
  type = NULL
)
}
\arguments{
\item{dt}{flow data.table}

\item{col}{character, column to compute net migration from,
typically flows or predictions.}

\item{by}{character, additional grouping variables besides o and
d. Typically agegroup and year.}

\item{o}{character, origin identifier}

\item{d}{character, destination identifier}

\item{long}{logical, should data be returned in long format?}

\item{type}{character, if specified a column called type is added
with entries as specified by type. Useful to identify
different tables}
}
\value{
data.table of net migration
}
\description{
Calculates net migration from flow data.table
}
\author{
Konstantin Hoffie
}
