% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_functions.R
\name{check_na_cols}
\alias{check_na_cols}
\title{Quickly check which columns of a dataset have NA values.}
\usage{
check_na_cols(data)
}
\arguments{
\item{data}{A data.frame, or other data that could contain NA values.}
}
\value{
Number of NA's in position of data.
}
\description{
Quickly check which columns of a dataset have NA values.
}
\details{
A function that gives a quick check of data to see where NA's are. Most useful for data.frames where the funciton will supply the number if NA's in each column. The output can be passed to the sum() function to check for total NA's.
}
\examples{
check_na_cols(iris)
}
