% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_gradebook.R
\name{clean_gradebook}
\alias{clean_gradebook}
\title{Clean gradebook}
\usage{
clean_gradebook(gb, cols)
}
\arguments{
\item{gb}{gradebook tibble}

\item{cols}{cols to clean}
}
\value{
cleaned gradebook
}
\description{
Given selected columns converts NA to 0 and EX to NA.
}
\examples{
clean_gradebook(gradebook, A1:A3)
}
