% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/string_functions.R
\name{collapse}
\alias{collapse}
\title{Collapse a series of strings}
\usage{
collapse(..., sep = "")
}
\arguments{
\item{...}{A vector, or vectors of strings}

\item{sep}{A delimeter on which two collapse strings}
}
\value{
A single string of collapsed input
}
\description{
Collapse a series of strings
}
\details{
A wrapper for paste0(..., collapse = <option>). Built to be more clear than paste0() when collapsing strings, and to avoid the issue of no partial string match in the paste() functions for the colllapse option.
}
\examples{
collapse(c("species", "yellow"), c("species", "yellow"), sep = ' ')
[1] "speciesspecies yellowyellow"
collapse(c("species", "yellow"))
[1] "speciesyellow"
}
