% Generated by roxygen2 (4.0.1): do not edit by hand
\name{combineDf}
\alias{combineDf}
\title{Combine data frames by matching specific columns.}
\usage{
combineDf(dfNew, df, commonCols = NULL, addCols = NULL,
  addNewRows = FALSE, suffixNewCols = ".new", verbose = TRUE)
}
\arguments{
\item{df1}{A data frame.}

\item{df2}{A data frame with the same columns as df1.}

\item{commonCols}{The columns that are present in both
data frames and which are used to identify corresponding
rows. If \code{NULL} the intersection of the column names
of the two data frames is used.}

\item{addCols}{Columns in dfNew to be added to df. If \code{NULL}
all columns but the \code{commonCols} are added.}

\item{addNewRows}{Add rows from \code{newDf} to \code{df}
even if there were no matches according to \code{commonCols}?}

\item{...}{Other atgumets that can be passed to \code{match}.}
}
\value{
For each row in df1 the (first) matching row in
df2, or \code{NA} otherwise.
}
\description{
Find matching columns of two data frames and
combine them to one data frame.
}
\examples{
\dontrun{
df = data.frame(a=0:3, b=1:4, c=2:5, d=letters[1:4])
dfNew <- cbind((df1+1)[3:1, 3:1], x=11:13, y=12:14)
dfComb <- combineDf(dfNew, df)
dfComb <- combineDf(dfNew, df, addNewRows=T)
}
}

