% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/string_functions.R
\name{concat}
\alias{concat}
\title{Concatenate a series of characters}
\usage{
concat(..., sep = "")
}
\arguments{
\item{...}{A vector, or vectors of strings}

\item{sep}{A delimeter on which to join strings}
}
\value{
A vector of character strings.
}
\description{
Concatenate a series of characters
}
\details{
A wrapper for paste(..., sep), called concat() to avoid confusion with collapse().
}
\examples{
concat("species", "yellow")
[1] "speciesyellow"
concat(c("species", "yellow"), c("species", "yellow"), sep = ' ')
[1] "species species" "yellow yellow"
}
