% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confmat2df.R
\name{confmat2df}
\alias{confmat2df}
\title{Convert a two-class confusion matrix in a data frame.}
\usage{
confmat2df(x, addCols = NULL, addColsfromAttr = NULL,
  stringsAsFactors = FALSE, long = T, key = "Measure",
  value = "Accuracy")
}
\arguments{
\item{x}{A two-class confusion matrix as returned by 
\code{\link{[caret]confusionMatrix}}.}

\item{addCols}{Add columns to the returned data frame. 
\code{as.data.frame(as.list(addCols))} must work and return 
a data frame with one row.}

\item{addColsfromAttr}{Name of the attribute to be used for 
\code{addCols}}

\item{stringsAsFactors}{...}

\item{long}{Return a data frame with the accuracies as 
key-value pairs}

\item{key}{If \code{long} is \code{TRUE} the name for the 
column with the accuracy measure names.}

\item{value}{If \code{long} is \code{TRUE} the name for the 
column with the accuracy values.}
}
\value{
\code{data.frame}
}
\description{
Convert a two-class confusion matrix in a data frame.
}

