% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_functions.R
\name{convert}
\alias{convert}
\title{Strict type conversion}
\usage{
convert(from, to)
}
\arguments{
\item{from}{Data object to be converted.}

\item{to}{Data object with class to be converted to}
}
\value{
A data object with the same class as `to` parameter, if possible.
}
\description{
Function to convert data from one class to that of another data object so that they are comparable.
}
\details{
Follows R's upwards coercion logic, and is not downwards compatible, i.e, a numeric can be converted to a string, but a string, even "1", will not be converted to numeric.
}
\examples{
convert(1, "2")
[1] "1"
convert(T, "I'm a test string")
[1] "TRUE"
}
