% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/convert_to_bin_confmat.R
\name{convert_to_bin_confmat}
\alias{convert_to_bin_confmat}
\title{Convert a confusion matrix in a two-cass confusion matrix.}
\usage{
convert_to_bin_confmat(cm, pos_class, newnames = c("Positive", "Negative"))
}
\arguments{
\item{cm}{A confsion matrix with. Can by a table or a ConfusionMatrix
from caret.}

\item{pos_class}{The name or column of the class which makes up the
first class of the new confusion matrix.
All other classes will be merged to one "other" class.}

\item{newnames}{Charater vector of length two with the classnames
of the returned two-class confusion matrix.}
}
\value{
A two-class table or ConfusionMatrix depending on the input.
}
\description{
...
}

