% Generated by roxygen2 (4.0.1): do not edit by hand
\name{create_clr}
\alias{create_clr}
\title{create_clr}
\usage{
create_clr(ids, labels, colors, ...)
}
\arguments{
\item{ids}{The class ids}

\item{labels}{The class labels}

\item{color}{The colors}

\item{...}{Not used.}
}
\value{
A color map
}
\description{
Create a colormap for ArcMAP.
}

