% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_package_helpeR.R
\name{create_package_helperR}
\alias{create_package_helperR}
\title{This function is used to transform a project into an installable R package.
The function wraps a series of functions from the \code{usethis} package to create
package.}
\usage{
create_package_helperR()
}
\description{
This function is used to transform a project into an installable R package.
The function wraps a series of functions from the \code{usethis} package to create
package.
}
\examples{
\dontrun{

#Run this in your console after creating a project
create_package_helperR()

}
}
