% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_semester.R
\name{create_semester}
\alias{create_semester}
\title{create semester}
\usage{
create_semester(start, weeks = 12, MSB = NULL)
}
\arguments{
\item{start}{date of start of semester}

\item{weeks}{number of weeks in semester}

\item{MSB}{week after which mid-semester breaks starts}
}
\value{
tibble with name and dates
}
\description{
Given a start date and number of weeks, this creates a tibble
with name of week and date. Can also add a two-week mid-semester
break (MSB)
}
\examples{
create_semester(lubridate::today(), weeks = 12, MSB = 8)
}
