% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createDataPartition_grouped.R
\name{createDataPartition_grouped}
\alias{createDataPartition_grouped}
\title{Title}
\usage{
createDataPartition_grouped(df, y, group, pTr = 0.5, seed = NULL)
}
\arguments{
\item{pTr}{}

\item{seed}{}

\item{ids}{}

\item{fids}{}
}
\description{
Title
}
\examples{
refset = data.frame(class=factor(rep(c(1, 2), c(4, 9))),
                    polygon=rep(1:5, c(2, 2, 2, 3, 4)),
                    x1=rnorm(13), x2=rnorm(13))

sets <- createDataPartition_grouped(refset, y='class', group='polygon', pTr=.5, seed=0)
sets
# The following check must be TRUE => pixels from a polygon are not found in both sets
length(intersect(sets$train$polygon, sets$test$polygon))==0
}

