% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_functions.R
\name{cross_tab}
\alias{cross_tab}
\title{Contingency table creation}
\usage{
cross_tab(var1, var2, margin = NULL)
}
\arguments{
\item{var1}{First object that can be passed to table()}

\item{var2}{Second object that can be passed to table()}

\item{margin}{Margin, as passed to prop.table. See ?prop.table for more details.}
}
\value{
Table object of the form form prop.table, but formatted to be a percentage to 1 decimal place.
}
\description{
Contingency table creation
}
\details{
Two variable cross tabulation of data. Output as a percentage to 1 decimal place.
}
\examples{
cross_tab(mtcars$disp, mtcars$mpg)
cross_tab_df(mtcars$disp, mtcars$mpg)
}
