% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/string_functions.R
\name{cs}
\alias{cs}
\title{Combine unquoted items into a vector of character strings}
\usage{
cs(...)
}
\arguments{
\item{...}{Unquoted single word strings.}
}
\value{
A vector of character strings.
}
\description{
Combine unquoted items into a vector of character strings
}
\details{
Works the same way that c() does for all types of data types. cs() allows you to quickly create vectors of single word strings without having to use quotes.
}
\examples{
cs(a,b,a)
[1] "a" "b" "a"
cs(Hello, green, 99)
[1] "Hello" "green" "99"
}
