% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tablemanip.R
\name{cuttable}
\alias{cuttable}
\title{Cut Table}
\usage{
cuttable(x, searchfor)
}
\arguments{
\item{x}{Data Frame to cut from}

\item{searchfor}{word in the prediction column to search for when cutting}
}
\description{
Cut Table
}
