% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mxDb.R
\name{dbAddGeoJSON}
\alias{dbAddGeoJSON}
\title{Add geojson list or file to db postgis}
\usage{
dbAddGeoJSON(geojsonList = NULL, geojsonPath = NULL, dbInfo = NULL,
  tableName = NULL, archiveIfExists = T, archivePrefix = "mx_archives")
}
\arguments{
\item{geojsonList}{list containing the geojson data}

\item{geojsonPath}{path the geojson}

\item{dbInfo}{dbInfo object containgin pass,user, ....}

\item{tableName}{Name of the postgis layer / table}
}
\description{
Add geojson list or file to db postgis
}

