% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mxDb.R
\name{dbGetColumnInfo}
\alias{dbGetColumnInfo}
\title{Get variable summary}
\usage{
dbGetColumnInfo(dbInfo = NULL, table = NULL, column = NULL)
}
\arguments{
\item{dbInfo}{Named list with dbName,host,port, user and password}

\item{table}{Table/layer from which extract extent}

\item{column}{Column/Variable on wich extract summary}
}
\description{
Get variable summary
}

