% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mxDb.R
\name{dbGetGeoJSON}
\alias{dbGetGeoJSON}
\title{Geojson from postGIS base}
\usage{
dbGetGeoJSON(dbInfo, query, fromSrid = "4326", toSrid = "4326",
  asList = FALSE)
}
\arguments{
\item{dbInfo}{Named list with dbName,host,port,user and password}

\item{query}{PostGIS spatial sql querry.}
}
\value{
geojson list
}
\description{
Geojson from postGIS base
}

