% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mxDb.R
\name{dbGetLayerExtent}
\alias{dbGetLayerExtent}
\title{Get layer extent}
\usage{
dbGetLayerExtent(dbInfo = NULL, table = NULL, geomColumn = "geom")
}
\arguments{
\item{dbInfo}{Named list with dbName,host,port,user and password}

\item{table}{Table/layer from which extract extent}

\item{geomColumn}{set geometry column}
}
\value{
extent
}
\description{
Get layer extent
}

