% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mxDb.R
\name{dbGetSp}
\alias{dbGetSp}
\title{Transfert postgis feature by sql query to sp object}
\usage{
dbGetSp(dbInfo, query)
}
\arguments{
\item{dbInfo}{Named list with dbName,host,port,user and password.}

\item{query}{PostGIS spatial sql querry.}
}
\value{
spatial object.
}
\description{
Transfert postgis feature by sql query to sp object
}

