% Generated by roxygen2 (4.0.1): do not edit by hand
\name{doOrReadRDS}
\alias{doOrReadRDS}
\title{doOrReadRDS}
\usage{
doOrReadRDS(expr, file, overwrite = FALSE)
}
\value{
The result of the expression (computed or loaded).
}
\description{
Evaluate an expression and write the result to disc or load it
if it already exists.
}

