% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/edit_vrt_color.R
\name{edit_vrt_color}
\alias{edit_vrt_color}
\title{Add a color table to a vrt file.}
\usage{
edit_vrt_color(id, col, file, backup = TRUE)
}
\arguments{
\item{id}{The IDs to be displayed with the colors given via \code{col}.}

\item{col}{The colors corresponding to the \code{id}.}

\item{file}{The vrt file to be edited.}

\item{backup}{Backup the vrt. Defaults to \code{TRUE}.}
}
\description{
It is assumed that the vrt represents a single band raster
having dataType="Byte". All values not given via id will be transparent.
}
\references{
#' The function has been inspired by
\url{https://gist.github.com/adammwilson/11052135}.
}
\seealso{
\code{\link{divColsAndBreaks}}, \code{\link{rgb24bit}}
}

