% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emptyRaster.R
\name{emptyRaster}
\alias{emptyRaster}
\title{emptyRaster}
\usage{
emptyRaster(x, ...)
}
\value{
A RasterLayer with the same number of rows/columns, crs and extent 
as the input Raster* object.
}
\description{
Create an empty raster layer given a Raster* object template.
}

