% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fill_grade_roster.R
\name{fill_grade_roster}
\alias{fill_grade_roster}
\title{fill grade roster}
\usage{
fill_grade_roster(df, infile, id_col = ID, RAA = 45:49)
}
\arguments{
\item{df}{scores with student_id col and total col}

\item{infile}{empty graderoster}

\item{id_col}{name of column with Ids}

\item{RAA}{marks to give RAA}
}
\value{
nothing
}
\description{
fill grade roster
}
