% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql_DB_helpers.R
\name{find_db_size}
\alias{find_db_size}
\title{This function will look inside a database table and counts the number of
rows inside that table.}
\usage{
find_db_size(conn, table_name)
}
\arguments{
\item{conn}{(\code{DBI} database connection object) This is the connection object
to a database. Refer to \code{DBI::dbConnect()} for details.}

\item{table_name}{(character) Name of the table in the database to extract
data from.}
}
\value{
(tibble) Returns a tibble that is randomly sampled from the chosen
table in the databse.
}
\description{
This function will look inside a database table and counts the number of
rows inside that table.
}
\examples{
\dontrun{

local_db_path <- "data/joined_lm_trig_data/mydb.db"
conn <- DBI::dbConnect(RSQLite::SQLite(), local_db_path)
tables <- RSQLite::dbListTables(conn) \%>\%
                 pluck(1)

sampled_data <-
        find_db_size(
        conn = conn,
        table_name = tables
        )

}
}
