% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitting.R
\name{fit_gravity}
\alias{fit_gravity}
\title{Fits simple gravity model}
\usage{
fit_gravity(dt, offset)
}
\arguments{
\item{dt}{data.table of flows}

\item{offset}{logical, if TRUE frompop and topop enter into model
as offset, meaning their coefficients are constrained to 1}
}
\value{
fit
}
\description{
Fits simple gravity model
}
\author{
Konstantin Hoffie
}
