% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_functions.R
\name{freq}
\alias{freq}
\title{Frequency table as data.frame}
\usage{
freq(..., sort = "none")
}
\arguments{
\item{...}{One or more objects that can be passed to table()}

\item{sort}{How the data should be sorted on frequency - ascending, descending or none.}
}
\value{
Sorted data.frame based either on first object, or by sort option object.
}
\description{
Frequency table as data.frame
}
\details{
Data.frame output of table with better column names and ability to sort on Freq column. Essentially creates crosstab table in long data format.
}
\examples{
freq(iris$Species)
freq(c(1,2,3,1,1,1,1))
}
