% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_nameUval_string.R
\name{get_nameUval_string}
\alias{get_nameUval_string}
\title{Title}
\usage{
get_nameUval_string(x, sep = ":", collapse = "|", collapse_vals = ",")
}
\arguments{
\item{x}{}

\item{sep}{}

\item{collapse}{}

\item{collapse_vals}{}
}
\description{
Title
}
\examples{
# With a list
x <- list(a=c(1:3), b=1)
get_nameUval_string(x)
get_nameUval_string(x, sep=" = ", collapse=" || ",  collapse_vals=",")

}

