% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mxVtDep.R
\name{glMakeUrl}
\alias{glMakeUrl}
\title{Create url for pgrestapi source}
\usage{
glMakeUrl(protocol = "http", host = "localhost", port, table,
  fieldVariables, fieldGeom)
}
\value{
url
}
\description{
Create url for pgrestapi source
}

