% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gradebook.R
\docType{data}
\name{gradebook}
\alias{gradebook}
\title{Example gradebook}
\format{
\subsection{\code{gradebook}}{

A data frame with 10 rows and 4 columns:
\describe{
\item{id}{Student ID}
\item{A1}{A1 marks}
\item{A2}{A2 marks}
\item{A3}{A3 marks}
}
}
}
\source{
simulated data
}
\usage{
gradebook
}
\description{
An example gradebook
Report ...
}
\keyword{datasets}
