% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_readers.R
\name{hread_table}
\alias{hread_table}
\title{Helper Read Table}
\usage{
hread_table()
}
\description{
An user-friendly and simplified way to use the read.table function for the lazy and those are just beggining.
Keep in mind that only the parameters path and header are available in this function.
}
