% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/number_functions.R
\name{integer_reverse}
\alias{integer_reverse}
\title{integer_reverse}
\usage{
integer_reverse(x)
}
\arguments{
\item{x}{A integer or numeric that can be coerced to an integer}
}
\value{
A rounded integer value
}
\description{
Reverse an integer in place
}
\details{
Reverses an integer x in place, for both positive and negative values. Does not work for vectors or non-integer data such as floats
}
\examples{
integer_reverse(123)
[1] 321
integer_reverse(-123)
[1] -321
}
