% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_functions.R
\name{is_}
\alias{is_}
\alias{is_zero}
\alias{is_natural}
\alias{is_decimal}
\title{Functions for boolean checks}
\usage{
is_zero(x, strict = F)

is_natural(x)

is_decimal(x)
}
\arguments{
\item{x}{A number/vector to pass to a boolean check}

\item{strict}{Whether a value is equal to the actual value, or also equal to natural coercion value}
}
\value{
A boolean, TRUE/FALSE
}
\description{
Functions to check for numerous different conditions
}
\details{
A list of is_* checks to be passed to other functions.
}
\section{Functions}{
\itemize{
\item \code{is_natural}: Check whether a number is a positive integer

\item \code{is_decimal}: Check to see if a number is a decimal, both positive and negative
}}

\examples{
is_zero(0, strict=T)
[1] TRUE
is_natural(4)
[1] TRUE
is_decimal(-8.17)
[1] TRUE
is_zero(FALSE, strict=T)
[1] FALSE
is_natural(4.4)
[1] FALSE
is_decimal(-8)
[1] FALSE
}
