% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_leap.R
\name{is_leap}
\alias{is_leap}
\title{is_leap}
\usage{
is_leap(year)
}
\arguments{
\item{year}{vector of year as input.}
}
\value{
a vector of true for leap years, and false if not.
}
\description{
Function that retun True for leap year.
}
\examples{
is_leap( year = seq( from = 2000, by = 1, lenght.out = 20) )
}
