% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jupyter_notebook.R
\name{jupyter_notebook}
\alias{jupyter_notebook}
\title{Activate specific conda environment and call jupyter notebook.}
\usage{
jupyter_notebook(env = "r-oneClass", bdir = getwd(), wait = FALSE,
  invisible = FALSE, ...)
}
\arguments{
\item{env}{The conda environment to be activated}

\item{bdir}{The directory where to launch jupyter}

\item{wait}{Changes default behaviour of the argument when running \code{\link{system}}}

\item{invisible}{Changes default behaviour of the argument when running \code{\link{system}}}

\item{...}{Other parameters that can be passed to to \code{\link{system}}}
}
\description{
Activate specific conda environment and call jupyter notebook.
}

