% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/land_kategorier.R
\name{land_kategorier_map}
\alias{land_kategorier_map}
\title{Oppslagstabel for landkoder og kategori}
\usage{
land_kategorier_map()
}
\value{
\code{data.frame} med \code{kode} (chr) og \code{kategori} (faktor).
}
\description{
Returnerer en data.frame med kolonnene \code{kode} og \code{kategori}
basert på listene over (nyttig for \code{left_join}).
}
\examples{
head(land_kategorier_map())
}
