% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mxMisc.R
\name{listToHtml}
\alias{listToHtml}
\title{R list to html}
\usage{
listToHtml(listInput, htL = "", h = 2, exclude = NULL)
}
\arguments{
\item{listInput}{list in inptu}

\item{htL}{List to append to}

\item{h}{Value of the first level of html header}

\item{exclude}{list named item to exclude}
}
\description{
R list to html
}

