% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mxMisc.R
\name{listToHtmlClass}
\alias{listToHtmlClass}
\title{R list to html list}
\usage{
listToHtmlClass(listInput, exclude = NULL, c = 0, htL = "",
  classUl = "list-group", classLi = "list-group-item")
}
\arguments{
\item{listInput}{list in inptu}

\item{exclude}{list named item to exclude}

\item{htL}{List to append to}

\item{h}{Value of the first level of html header}
}
\value{
HTML list
}
\description{
Create a html list and apply a class for <ul> and <li>
}

