% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_custom_functions.R
\name{load_custom_functions}
\alias{load_custom_functions}
\title{This function goes through a specified folder within a project and sources
all the .R files within the folder. This functions like
\code{devtools::load_all()} for non-package repos. Please ensure the contents of
.R files within the folder are functions/variables that you would like in
your global environment before running the funtion.}
\usage{
load_custom_functions(.path = "R")
}
\arguments{
\item{.path}{(character) The file path of the folder you would like the .R
files sourced from.}
}
\description{
This function goes through a specified folder within a project and sources
all the .R files within the folder. This functions like
\code{devtools::load_all()} for non-package repos. Please ensure the contents of
.R files within the folder are functions/variables that you would like in
your global environment before running the funtion.
}
\examples{
\dontrun{

load_custom_functions(.path = "R")

}
}
