% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_functions.R
\name{lunique}
\alias{lunique}
\title{The number of unique elements in an object}
\usage{
lunique(x)
}
\value{
The number of unique elements
}
\description{
Function to calculate the length of the numebr of unique elements in a data object.
}
\details{
Just a super convenient wrapper for the length(unique(x)) combination. See ?unique for more details.
}
\examples{
lunique(iris$Species)
[1] 3

sapply(iris, lunique)
Sepal.Length  Sepal.Width Petal.Length  Petal.Width      Species
35           23           43           22            3
}
