% Generated by roxygen2 (4.0.1): do not edit by hand
\name{matchDf}
\alias{matchDf}
\title{matchDf}
\usage{
matchDf(df1, df2, ...)
}
\arguments{
\item{df1}{A data frame.}

\item{df2}{A data frame with the same columns as df1.}

\item{...}{Other atgumets that can be passed to \code{match}.}
}
\value{
For each row in df1 the (first) matching row in
df2, or \code{NA} otherwise.
}
\description{
Find the positions of (first) matches of the
first argument in the second.
}
\examples{
\dontrun{
df1 = data.frame(a=1:3, b=2:4, c=3:5)
df2 <- cbind((df1+1)[3:1, 3:1], x=11:13, y=12:14)
matches = matchDf(df1, df2)
df2[!is.na(matches), ]
}
}

