% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MC_anocAIC.R
\name{MC_anocAIC}
\alias{MC_anocAIC}
\title{Multi-Core Version von Comparison of several Models via cAIC}
\usage{
MC_anocAIC(objects, digits = 2)
}
\arguments{
\item{objects}{list() an Kandidaten Modelle}

\item{digits}{Anzahl der auszugebene Nachkommastelle}
}
\description{
Unterschied zur originalen Version, ist dass hier Multi-Core Support besteht.
Desweiteren ist es möglich mit dieser Version auch lm-Objekte in der Liste an
Kandidaten Modellen zu nutzen.
}
\references{
Saefken, B., Ruegamer, D., Kneib, T. and Greven, S. (2018):
Conditional Model Selection in Mixed-Effects Models with cAIC4.
\url{https://arxiv.org/abs/1803.05664}

Greven, S. and Kneib T. (2010) On the behaviour of marginal and
conditional AIC in linear mixed models. Biometrika 97(4), 773-789.
}
\seealso{
\code{\link[cAIC4]{cAIC4-package}}, \code{\link[lme4]{lmer}}
}
\author{
Rene-Marcel Kruse
}
\keyword{cAIC}
\keyword{multicore,}
