% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MC_getWeights.R
\name{MC_getWeights}
\alias{MC_getWeights}
\title{Optimize weights for model averaging.}
\usage{
MC_getWeights(models)
}
\arguments{
\item{models}{An list object containing all considered candidate models fitted by
\code{\link[lme4]{lmer}} of the lme4-package or of class
\code{\link[nlme]{lme}}.}
}
\value{
An updated object containing a vector of weights for the underlying candidate models, value
of the object given said weights as well as the time needed.
}
\description{
Function to constructed an optimal vector of weights for model averaging of
Linear Mixed Models based on the proposal of Zhang et al. (2014) of using Stein's Formular
to derive a suitable criterion based on the conditional Akaike Information Criterion as
proposed by Greven and Kneib. The underlying optimization used is a customized version
of the Augmented Lagrangian Method.
}
\author{
Rene-Marcel Kruse
}
\keyword{augmented}
\keyword{averaging,}
\keyword{lagrangian}
\keyword{method}
\keyword{model}
\keyword{weights,}
