% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MC_modelAvg.R
\name{MC_modelAvg}
\alias{MC_modelAvg}
\title{Multi-Core Version von Model Averaging Funktionen aus cAIC4-Package}
\usage{
MC_modelAvg(models, opt = TRUE)
}
\arguments{
\item{models}{liste an kandidaten modellen fürs model averaging}

\item{opt}{Parameter ob die Augmented Lagrangian Optimierung genutzt werden soll. TRUE = JA.}
}
\description{
Unterschied zur originalen Version, ist dass hier Multi-Core Support besteht.
}
\references{
Saefken, B., Ruegamer, D., Kneib, T. and Greven, S. (2018):
Conditional Model Selection in Mixed-Effects Models with cAIC4.
\url{https://arxiv.org/abs/1803.05664}

Greven, S. and Kneib T. (2010) On the behaviour of marginal and
conditional AIC in linear mixed models. Biometrika 97(4), 773-789.
}
\seealso{
\code{\link[cAIC4]{cAIC4-package}}
}
\author{
Rene-Marcel Kruse
}
\keyword{averaging}
\keyword{cAIC,}
\keyword{model}
\keyword{multicore,}
