% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_functions.R
\name{mod0}
\alias{mod0}
\title{Special case modulus functions}
\usage{
mod0(x, y)
}
\arguments{
\item{x}{A numeric value}

\item{y}{A base for modulus function}
}
\value{
The remainder between x and y, x mod y
}
\description{
Alternative mod function which returns the original value if base is 0, passed to is_decimal & is_natural functions
}
\details{
A work around where you need modulus function to return the input value, not NaN
}
\examples{
mod0(5, 2)
[1] 1
mod0(5, 0)
[1] 5
}
