% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mxUi.R
\name{mxAccordionGroup}
\alias{mxAccordionGroup}
\title{Create a bootstrap accordion}
\usage{
mxAccordionGroup(id, style = NULL, show = NULL, itemList)
}
\arguments{
\item{id}{Accordion group ID}

\item{style}{Additional style.}

\item{show}{Vector of item number. Collapse all item except those in this list. E.g. c(1,5) will open items 1 and 5 by default.}

\item{itemList}{Nested named list of items, containing title and content items. E.g. list("foo"=list("title"="foo","content"="bar"))}
}
\description{
Create a bootstrap accordion element, based on a named list.
}
\examples{
amAccordionGroup(id='superTest',
 itemList=list(
   'a'=list('title'='superTitle',content='acontent'),
   'b'=list('title'='bTitle',content='bContent'))
 )
}

