% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/mapx.R
\name{mxAllow}
\alias{mxAllow}
\title{Control ui access}
\usage{
mxAllow(logged, roleName, roleLowerLimit)
}
\arguments{
\item{logged}{Boolean. Is the user logged in ?}

\item{roleName}{Character. Role in numeric format}

\item{roleLowerLimit}{Numeric. Minumum role requirement}
}
\description{
Use mxConfig$roleVal list to check if the curent user's role name can access to the given numeric role.
}

