% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mxMisc.R
\name{mxAnalysisOverlaps}
\alias{mxAnalysisOverlaps}
\title{Overlaps analysis}
\usage{
mxAnalysisOverlaps(inputBaseLayer, inputMaskLayer, outName,
  dataOwner = "mapxw", sridOut = 4326, varToKeep = "gid")
}
\description{
Use a mask to get overlaps over a layer
}

