% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mxMisc.R
\name{mxCanReach}
\alias{mxCanReach}
\title{Test for internet connection. 
The idea is to reach google with a ping and determine if there is a full packet response without loss}
\usage{
mxCanReach(server = "google.com", port = 80)
}
\arguments{
\item{host}{String. Host name to ping}
}
\description{
Test for internet connection. 
The idea is to reach google with a ping and determine if there is a full packet response without loss
}

