% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/mapx.R
\name{mxCatch}
\alias{mxCatch}
\title{Catch errors}
\usage{
mxCatch(title, expression, session = shiny:::getDefaultReactiveDomain(),
  debug = TRUE, panelId = "panelAlert")
}
\arguments{
\item{title}{Title of the alert}

\item{session}{Shiny session object}

\item{debug}{Boolean. Return also message as alert.}

\item{panelId}{Id of the output element}
}
\description{
Catch errors and return alert panel in an existing div id.
}

