% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mxMisc.R
\name{mxCheckboxIcon}
\alias{mxCheckboxIcon}
\title{Set a checkbox button with custom icon.}
\usage{
mxCheckboxIcon(id, idLabel, icon, display = TRUE)
}
\arguments{
\item{id}{Id of the element}

\item{icon}{Name of the fontawesome icon. E.g. cog, times, wrench}
}
\description{
Create a checkbox input with a select icon.
}

