% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mxDb.R
\name{mxDbAddGeoJSON}
\alias{mxDbAddGeoJSON}
\title{Add geojson list or file to db postgis}
\usage{
mxDbAddGeoJSON(geojsonList = NULL, geojsonPath = NULL, tableName = NULL,
  archiveIfExists = T, archivePrefix = "mx_archives")
}
\arguments{
\item{geojsonList}{list containing the geojson data}

\item{geojsonPath}{path the geojson}

\item{tableName}{Name of the postgis layer / table}
}
\description{
Add geojson list or file to db postgis
}

