% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mxDb.R
\name{mxDbClearAll}
\alias{mxDbClearAll}
\title{Remove old results from db query}
\usage{
mxDbClearAll(dbInfo)
}
\arguments{
\item{dbInfo}{Named list with dbName,host,port,user and password}
}
\description{
Remove old results from db query
}

