% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mxDb.R
\name{mxDbDropLayer}
\alias{mxDbDropLayer}
\title{drop layer
layerName Layer (table + entry + views) to delete from db}
\usage{
mxDbDropLayer(layerName)
}
\description{
drop layer
layerName Layer (table + entry + views) to delete from db
}

