% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mxMisc.R
\name{mxDbEncrypt}
\alias{mxDbDecrypt}
\alias{mxDbEncrypt}
\title{Encrypt or decrypt data using postgres pg_sym_encrypt}
\usage{
mxDbEncrypt(data, ungroup = FALSE, key = mxConfig$key)

mxDbDecrypt(data = NULL, key = mxConfig$key)
}
\arguments{
\item{data}{vector, list or data.frame to encrypt or decrypt}

\item{ungroup}{boolean : ungroup the data and apply the encryption on individual item.}

\item{key}{Encryption key}
}
\value{
encrypted data as list
}
\description{
Encrypt or decrypt data using postgres pg_sym_encrypt
}

