% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mxMisc.R
\name{mxDbExistsTable}
\alias{mxDbExistsTable}
\title{Check if table exists in postgresql}
\usage{
mxDbExistsTable(dbInfo, table)
}
\arguments{
\item{dbInfo}{Named list with dbName,host,port,user and password}

\item{table}{Name of the table to check}
}
\description{
Shortcut to create a connection, and check if table exists.
}

