% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mxDb.R
\name{mxDbGetColumnInfo}
\alias{mxDbGetColumnInfo}
\title{Get variable summary}
\usage{
mxDbGetColumnInfo(table = NULL, column = NULL)
}
\arguments{
\item{table}{Table/layer from which extract extent}

\item{column}{Column/Variable on wich extract summary}
}
\description{
Get variable summary
}

