% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mxDb.R
\name{mxDbGetColumnsNames}
\alias{mxDbGetColumnsNames}
\title{List existing column from postgresql table}
\usage{
mxDbGetColumnsNames(table)
}
\arguments{
\item{dbInfo}{Named list with dbName,host,port,user and password}
}
\description{
Shortcut to get column name for a table
}

