% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mxDb.R
\name{mxDbGetColumnsTypes}
\alias{mxDbGetColumnsTypes}
\title{List existing column type from postgresql table}
\usage{
mxDbGetColumnsTypes(table)
}
\arguments{
\item{table}{Name of the table to evaluate}
}
\description{
Shortcut to get column type for a table
}

