% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mxDb.R
\name{mxDbGetGeoJSON}
\alias{mxDbGetGeoJSON}
\title{Geojson from postGIS base}
\usage{
mxDbGetGeoJSON(query, fromSrid = "4326", toSrid = "4326", asList = FALSE)
}
\arguments{
\item{query}{PostGIS spatial sql querry.}
}
\value{
geojson list
}
\description{
Geojson from postGIS base
}

