% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mxDb.R
\name{mxDbGetLayerCentroid}
\alias{mxDbGetLayerCentroid}
\title{Get layer center}
\usage{
mxDbGetLayerCentroid(table = NULL, geomColumn = "geom")
}
\arguments{
\item{table}{Table/layer from which extract extent}

\item{geomColumn}{set geometry column}
}
\value{
extent
}
\description{
Compute the union of all geometry in a given layer and return the coordinate of the centroid.
}

