% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mxDb.R
\name{mxDbGetLayerExtent}
\alias{mxDbGetLayerExtent}
\title{Get layer extent}
\usage{
mxDbGetLayerExtent(table = NULL, geomColumn = "geom")
}
\arguments{
\item{table}{Table/layer from which extract extent}

\item{geomColumn}{set geometry column}
}
\value{
extent
}
\description{
Get layer extent
}

