% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mxDb.R
\name{mxDbGetSp}
\alias{mxDbGetSp}
\title{Transfert postgis feature by sql query to sp object}
\usage{
mxDbGetSp(query)
}
\arguments{
\item{query}{PostGIS spatial sql querry.}
}
\value{
spatial object.
}
\description{
Transfert postgis feature by sql query to sp object
}

