% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mxDb.R
\name{mxDbGetUserInfoList}
\alias{mxDbGetUserInfoList}
\title{Get user info}
\usage{
mxDbGetUserInfoList(id = NULL, email = NULL, userTable = "mx_users")
}
\arguments{
\item{email}{user email}

\item{userTable}{DB users table}
}
\value{
list containing data from the user
}
\description{
Get user info
}

