% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mxMisc.R
\name{mxDbGetUsersGroups}
\alias{mxDbGetUsersGroups}
\title{Get group table for users}
\usage{
mxDbGetUsersGroups(idFilter = NULL)
}
\arguments{
\item{idFilter}{optional filter of vector containing ids}
}
\description{
Get group table for users
}

