% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mxDb.R
\name{mxDbListColumns}
\alias{mxDbListColumns}
\title{List existing column from postgresql table}
\usage{
mxDbListColumns(dbInfo, table)
}
\arguments{
\item{dbInfo}{Named list with dbName,host,port,user and password}
}
\description{
Shortcut to create a connection, get the list of column and close the connection, using a dbInfo list.
}

