% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mxDb.R
\name{mxDbUpdate}
\alias{mxDbUpdate}
\title{Update a single value of a table}
\usage{
mxDbUpdate(table, column, idCol = "id", id, value, expectedRowsAffected = 1)
}
\arguments{
\item{table}{Table to update}

\item{column}{Column to update}

\item{idCol}{Column of identification}

\item{id}{Identification value}

\item{value}{Replacement value}

\item{expectedRowsAffected}{Number of row expected to be affected. If the update change a different number of row than expected, the function will rollback}
}
\value{
Boolean worked or not
}
\description{
Update a single value of a table
}

