% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mxDb.R
\name{mxDbWriteSpatial}
\alias{mxDbWriteSpatial}
\title{Write spatial data frame to postgis}
\usage{
mxDbWriteSpatial(spatial.df = NULL, schemaname = "public", tablename,
  overwrite = FALSE, keyCol = "gid", srid = 4326, geomCol = "geom")
}
\arguments{
\item{spatial.df}{Spatial  data frame object}

\item{schemaname}{Target schema table}

\item{tablename}{Target table name}

\item{overwrite}{Overwrite if exists}

\item{keyCol}{Set new primary key}

\item{srid}{Set the epsg code / SRID}

\item{geomCol}{Set the name of the geometry column}
}
\description{
Convert spatial data.frame to postgis table. Taken from https://philipphunziker.wordpress.com/2014/07/20/transferring-vector-data-between-postgis-and-r/
}

