% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/mapx.R
\name{mxDebugMsg}
\alias{mxDebugMsg}
\title{Print debug message}
\usage{
mxDebugMsg(m = "")
}
\arguments{
\item{m}{Message to be printed}
}
\description{
Print a defaut debug message with date as prefix. NOTE: this function should take a global parameter "debug" and a log file.
}

