% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mxMisc.R
\name{mxEitiGetCountryCenter}
\alias{mxEitiGetCountryCenter}
\title{Create a formated list of country center from eiti countries table}
\usage{
mxEitiGetCountryCenter(eitiCountryTable)
}
\description{
Create a formated list of country center from eiti countries table
}

