% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mxMisc.R
\name{mxEitiGetCountrySelectizeList}
\alias{mxEitiGetCountrySelectizeList}
\title{Create a formated list for selectize input from eiti countries table}
\usage{
mxEitiGetCountrySelectizeList(eitiCountryTable)
}
\description{
Create a formated list for selectize input from eiti countries table
}

