% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mxMisc.R
\name{mxEmailIsKnown}
\alias{mxEmailIsKnown}
\title{Check if an email is known and active}
\usage{
mxEmailIsKnown(email = NULL, usertable = "mx_users", active = TRUE,
  validated = TRUE)
}
\arguments{
\item{email}{map-x user email}

\item{usertable}{name of the table}
}
\value{
boolean exists
}
\description{
Check in a standard mapx database if an email/user exists
}

