% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mxMisc.R
\name{mxEmailIsValid}
\alias{mxEmailIsValid}
\title{Check if given email is valid}
\usage{
mxEmailIsValid(email = NULL)
}
\arguments{
\item{email}{String email address to verify}
}
\value{
named logic vector
}
\description{
Check if given email is valid
}

