% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mxUi.R
\name{mxFileInput}
\alias{mxFileInput}
\title{Custom file input}
\usage{
mxFileInput(inputId, label, fileAccept = NULL, multiple = FALSE)
}
\arguments{
\item{inputId}{id of the file input}

\item{label}{Label for the input}

\item{fileAccept}{List of accepted file type. Could be extension.}

\item{multiple}{Boolean. Allow multiple file to be choosen. Doesn't work on all client.}
}
\description{
Default shiny fileInput has no option for customisation. This function allows to fully customize file input using the label tag.
}

