% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mxMisc.R
\name{mxGetLayerMeta}
\alias{mxGetLayerMeta}
\title{Get layer meta stored in default layer table}
\usage{
mxGetLayerMeta(dbInfo, layer)
}
\arguments{
\item{dbInfo}{Named list with dbName,host,port,user and password}

\item{layer}{Postgis layer stored in layer table. Should have a meta field.}
}
\description{
Get layer meta stored in default layer table
}

