% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mxMisc.R
\name{mxGetListValue}
\alias{mxGetListValue}
\title{Extract value from a list using path}
\usage{
mxGetListValue(li, path)
}
\arguments{
\item{li}{Input named list}

\item{path}{Path inside the list}
}
\value{
value extracted or NULL
}
\description{
Extract value from a list using path
}

