% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mxMisc.R
\name{mxGetMaxRole}
\alias{mxGetMaxRole}
\title{Return the highest role for a given user}
\usage{
mxGetMaxRole(project, userInfo, useWorld = T)
}
\arguments{
\item{project}{Project to look for}

\item{userInfo}{object of class mxUserInfoList produced with mxDbGetUserInfoList}

\item{useWorld}{Boolean keep result for project "world" in the result}
}
\description{
Return the highest role for a given user
}

