% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mxStoryMap.R
\name{mxGetStoryMapData}
\alias{mxGetStoryMapData}
\title{Get story map data}
\usage{
mxGetStoryMapData(id)
}
\arguments{
\item{id}{Id of the story map to get}
}
\value{
Story map content and visibility
}
\description{
Get story map data
}

