% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/mapx.R
\name{mxGetViewsTable}
\alias{mxGetViewsTable}
\title{Retrieve map views table}
\usage{
mxGetViewsTable(dbInfo = NULL, table = "mx_views", validated = TRUE,
  archived = FALSE, country = "AFG")
}
\arguments{
\item{dbInfo}{Named list with dbName,host,port, user and password}

\item{table}{Table name containing views info}

\item{validated}{Boolean filter validated dataset. Default = TRUE}

\item{archived}{Boolean filter to get archived data. Default =FALSE}

\item{country}{ISO 3 code to filter country.}
}
\description{
Get a list of available map-x views in given table, e.g. mx_views
}

