% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mxMisc.R
\name{mxMakeViewList}
\alias{mxMakeViewList}
\title{extract views from the db and create a list}
\usage{
mxMakeViewList(country, visibility, userId)
}
\arguments{
\item{cntry}{Country iso3 code}
}
\value{
list of views data and style
}
\description{
extract views from the db and create a list
}

