% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mxUi.R
\name{mxMakeViews}
\alias{mxMakeViews}
\title{Create html list of available views}
\usage{
mxMakeViews(views)
}
\arguments{
\item{views}{List of available views}
}
\description{
get a list of views and return a HTML shiny checkbox input.
}

