% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/mapx.R
\name{mxPanel}
\alias{mxPanel}
\title{Create a modal panel}
\usage{
mxPanel(id = "default", title = NULL, subtitle = NULL, html = NULL,
  listActionButton = NULL, background = TRUE, defaultButtonText = "OK",
  style = NULL, class = NULL, hideCloseButton = FALSE, draggable = TRUE)
}
\arguments{
\item{id}{Panel id}

\item{title}{Panel title}

\item{subtitle}{Panel subtitle}

\item{html}{HTML content of the panel, main text}

\item{listActionButton}{If FALSE, hide buttons. If NULL, display default close panel button, with text given in defaultButtonText. If list of buttons, list of button.}

\item{defaultButtonText}{Text of the default button if listActionButton is NULL and not FALSE}

\item{style}{Additional CSS style for the panel}

\item{class}{Additional class for the panel}

\item{hideCloseButton}{Boolean. Hide the close panel button}

\item{draggable}{Boolean. Set the panel as draggable}
}
\description{
Create a modal panel with some options as custom button, close button, html content.
}

