% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mxUi.R
\name{mxPanelAlert}
\alias{mxPanelAlert}
\title{Alert panel}
\usage{
mxPanelAlert(title = c("error", "warning", "message"), subtitle = NULL,
  message = NULL, listActionButton = NULL, ...)
}
\arguments{
\item{title}{Title of the alert. Should be "error", "warning" or "message"}

\item{subtitle}{Subtitle of the alert}

\item{message}{html or text message for the alert}

\item{listActionButtons}{List of action button for the panel}
}
\description{
Create an alert panel. This panel could be send to an output object from a reactive context.
}

