% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mxMisc.R
\name{mxParseListFromText}
\alias{mxParseListFromText}
\title{Parse key value pair from text}
\usage{
mxParseListFromText(txt)
}
\arguments{
\item{txt}{unformated text with key value pair. eg. myKey = myValue}
}
\value{
list of value
}
\description{
Parse key value pair from text
}

