% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mxStoryMap.R
\name{mxParseStory}
\alias{mxParseStory}
\title{Parse story map : markdown, R, view and video}
\usage{
mxParseStory(txtorig, knit = T, toc = F)
}
\arguments{
\item{test}{Story map text}
}
\value{
parsed html
}
\description{
Parse story map : markdown, R, view and video
}

