% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mxMisc.R
\name{mxParseView}
\alias{mxParseView}
\title{Parse view string}
\usage{
mxParseView(text)
}
\arguments{
\item{test}{Story map text with @view_start( name ; id ; extent ) ... @view_end tags}
}
\value{
parsed html
}
\description{
Parse view string
}

