% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mxMisc.R
\name{mxParseVimeo}
\alias{mxParseVimeo}
\title{Parse vimeo string}
\usage{
mxParseVimeo(text)
}
\arguments{
\item{text}{Story map text with @vimeo( id ; desc ) tag}
}
\value{
html enabled version
}
\description{
Parse vimeo string
}

