% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mxMisc.R
\name{mxRecursiveSearch}
\alias{mxRecursiveSearch}
\title{Recursive search and filter on named list}
\usage{
mxRecursiveSearch(li, column = "", operator = "==", search = "",
  filter = "")
}
\arguments{
\item{li}{List to evaluate}

\item{column}{Named field to search on (unique)}

\item{operator}{Search operator ('>','<','==','>=','<=','!=','%in%')}

\item{search}{Value to search}

\item{filter}{Named field to keep}
}
\value{
list or named vector if filter is given
}
\description{
Recursive search and filter on named list
}

