% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mxMisc.R
\name{mxRemoveEl}
\alias{mxRemoveEl}
\title{remove element by class or id}
\usage{
mxRemoveEl(session = getDefaultReactiveDomain(), class = NULL, id = NULL)
}
\arguments{
\item{session}{default shiny session}

\item{class}{class name to remove}

\item{id}{id to remove}
}
\description{
remove element by class or id
}

