% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mxMisc.R
\name{mxSendJson}
\alias{mxSendJson}
\title{function to read json and save as an object}
\usage{
mxSendJson(pathToJson, objName, session = getDefaultReactiveDomain())
}
\description{
function to read json and save as an object
}

