% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mxMisc.R
\name{mxSendMail}
\alias{mxSendMail}
\title{Send an email using local or remote 'mail' command}
\usage{
mxSendMail(from = mxConfig$mapxBotEmail, to = NULL, body = "",
  subject = "", wait = FALSE)
}
\arguments{
\item{from}{String. Valid email for  sender}

\item{to}{String. Valid email for Recipient}

\item{body}{String. Text of the body}

\item{subject.}{String. Test for the subject}
}
\description{
Send an email using local or remote 'mail' command
}

