% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mxMisc.R
\name{mxSetCookie}
\alias{mxSetCookie}
\title{Save named list of value into cookie}
\usage{
mxSetCookie(cookie = NULL, expireDays = NULL, deleteAll = FALSE,
  reloadPage = FALSE, session = getDefaultReactiveDomain())
}
\arguments{
\item{cookie}{Named list holding paired cookie value. e.g. (list(whoAteTheCat="Alf"))}

\item{expireDays}{Integer of days for the cookie expiration}

\item{session}{Shiny session object. By default: default reactive domain.}

\item{read}{Boolean. Read written cookie}
}
\description{
Note : don't use this for storing sensitive data, unless you have a trusted network.
}

