% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/mapx.R
\name{mxSetMapPanelMode}
\alias{mxSetMapPanelMode}
\title{Set map panel mode.}
\usage{
mxSetMapPanelMode(session = shiny::getDefaultReactiveDomain(),
  mode = c("mapViewsConfig", "mapViewsCreator", "mapStoryCreator",
  "mapViewsExplorer", "mapViewsToolbox"), title = NULL)
}
\arguments{
\item{session}{Shiny session}

\item{mode}{Map panel mode. In mapViewCreator, mapStoryCreator, mapExplorer}

\item{title}{Optionnal title to be returned.}
}
\value{
title string
}
\description{
Map-x panel use multiple panel mode : config,creator,explorer,toolbox. This function set and save the panel mode.
}

