% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mxMisc.R
\name{mxSetStyle}
\alias{mxSetStyle}
\title{Apply map-x style to existing vector tiles}
\usage{
mxSetStyle(session = shiny:::getDefaultReactiveDomain(), style,
  mapId = "mapxMap")
}
\arguments{
\item{session}{Shiny session object}

\item{style}{map-x style}
}
\description{
When leafletvt handle a vector tile source, a lealflet object is stored in leafletvtId, but no style is applied. Default is transparent. We add a style function after that the layer is fully loaded using this function. The style function is also stored alongside the leaflet object in leafletId under the name "vtStyle".
}

