% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/mapx.R
\name{mxSliderOpacity}
\alias{mxSliderOpacity}
\title{Set ioRange slider for opacity}
\usage{
mxSliderOpacity(id, opacity)
}
\arguments{
\item{id}{Id of the slider}

\item{opacity}{Default opacity}
}
\description{
Set ioRange slider for opacity
}

