% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mxMisc.R
\name{mxTextValidation}
\alias{mxTextValidation}
\title{Sting validation}
\usage{
mxTextValidation(textToTest, existingTexts, idTextValidation, minChar = 5,
  testForDuplicate = TRUE, testForMinChar = TRUE,
  displayNameInValidation = TRUE, existsText = "taken",
  errorColor = "#FF0000")
}
\arguments{
\item{existingTexts}{Vector of existing text}

\item{idTextValidation}{Id of the ui element to update (id=example -> uiOutput("example"))}

\item{minChar}{Minimum character length}

\item{testForDuplicate}{Boolean test for duplicate.}

\item{testForMinChar}{Boolean test for minimum number of character}

\item{displayNameInValidation}{Boolean add text in validation text}

\item{textTotest}{text to test against rules}
}
\value{
boolean : valid or not
}
\description{
Check if a string exists in a vector of string, if there is a duplicate, if contain at least n character, etc.. and update an existing div with a html summary. Return if the string is valid or not.
}

