% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mxUi.R
\name{mxTimeSlider}
\alias{mxTimeSlider}
\title{Set ioRange slider for time slider}
\usage{
mxTimeSlider(id, min, max, lay)
}
\arguments{
\item{id}{Id of the slider}

\item{min}{Minimum js unix date in milisecond}

\item{max}{Maxmimum js unix date in milisecond}

\item{lay}{Layer name}
}
\description{
Return a div than contain a slider input instantiated with ionRangeSlider for view time slider.
}

