% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mxMisc.R
\name{mxUiAccess}
\alias{mxUiAccess}
\title{Control ui access
 
UI  manager based on login info}
\usage{
mxUiAccess(logged, roleNum, roleLowerLimit, uiDefault, uiRestricted)
}
\arguments{
\item{logged}{Boolean. Is the user logged in ?}

\item{roleNum}{Numeric. Role in numeric format}

\item{roleLowerLimit}{Numeric. Minumum role requirement}

\item{uiDefault}{TagList. Default ui.}

\item{uiRestricted}{TagList. Restricted ui.}
}
\description{
Control ui access
 
UI  manager based on login info
}

