% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mxMisc.R
\name{mxUiEnable}
\alias{mxUiEnable}
\title{Control visbility of elements}
\usage{
mxUiEnable(session = shiny:::getDefaultReactiveDomain(), id = NULL,
  class = NULL, enable = TRUE, classToRemove = "mx-hide")
}
\arguments{
\item{session}{Shiny session}

\item{id}{Id of element to enable/disable}

\item{enable}{Boolean. Enable or not.}
}
\description{
Display or hide element by id, without removing element AND without having element's space empty in UI. This function add or remove mx-hide class to the element.
}

