% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mxMisc.R
\name{mxUpdateChartRadar}
\alias{mxUpdateChartRadar}
\title{Create a chartRadar in a canvas element.}
\usage{
mxUpdateChartRadar(session = shiny::getDefaultReactiveDomain(), main,
  compMain, id, idLegend, labels, values, compValues)
}
\arguments{
\item{session}{Shiny reactive session}

\item{main}{Main label}

\item{compMain}{Comparative value label}

\item{id}{Id of the canvas}

\item{idLegend}{Id of the legend}

\item{labels}{Labels for value and comparative values}

\item{compValues}{Comparative values}

\item{value}{Values}
}
\description{
Search the dom for an id a get drawing context, create a new chart object and config it with data.
}

