% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mxMisc.R
\name{mxUpdatePanel}
\alias{mxUpdatePanel}
\title{Update existing panel}
\usage{
mxUpdatePanel(panelId = NULL, session = shiny:::getDefaultReactiveDomain(),
  ...)
}
\arguments{
\item{panelId}{Id of the existing panel}

\item{session}{Shiny reactive object of the session}

\item{...}{Other mxPanel options}
}
\description{
Use output object to update the panel with a known id. E.g. for updating uiOutput("panelTest"), use mxUpdatePanel with panelId "panelTest"
}

