% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mxMisc.R
\name{mxUpdateText}
\alias{mxUpdateText}
\title{Update text by id}
\usage{
mxUpdateText(id, text = NULL, ui = NULL, addId = FALSE,
  session = shiny:::getDefaultReactiveDomain())
}
\arguments{
\item{id}{Id of the element}

\item{text}{New text}

\item{session}{Shiny session}
}
\description{
Search for given id and update content.
}

