% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{neg_clean}
\alias{neg_clean}
\title{Negative Cleaner}
\usage{
neg_clean(x)
}
\arguments{
\item{x}{The Data Frame user wishes to process.}
}
\description{
Substitutes all the negative values in the given Data Frame for 0.
}
