% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{negclean}
\alias{negclean}
\title{Negative Cleaner}
\usage{
negclean(x)
}
\arguments{
\item{x}{The Data Frame user wishes to process.}
}
\description{
Substitutes all the negative values in the given Data Frame for 0.
}
