% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tablemanip.R
\name{neural.arrange}
\alias{neural.arrange}
\title{Neural Arrange}
\usage{
neural.arrange(x)
}
\arguments{
\item{x}{Data frame to be arranged}
}
\description{
Organizes results from neural nework predictions. Creates a column and inserts in each row the name of the column with the biggest probability to occur. By the end, appends this columns to the inserted dataframe.
}
