% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mxMisc.R
\name{noDataCheck}
\alias{noDataCheck}
\title{Check for no null, NA's, nchar of 0, lenght of 0  or "[NO DATA]" string in a vector.}
\usage{
noDataCheck(val, useNoData = TRUE, noDataVal = "[ NO DATA ]")
}
\arguments{
\item{val}{Vector to test for no data.}
}
\value{
TRUE if no data (nchar == 0 OR is.na OR is.null) found or if input is not a vector
}
\description{
Check for no null, NA's, nchar of 0, lenght of 0  or "[NO DATA]" string in a vector.
}

