% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infix_functions.R
\name{not_in}
\alias{not_in}
\alias{\%not in\%}
\title{Value Matching}
\usage{
... \%not in\% NA
}
\value{
TRUE or FALSE value whether an element is in the object or not.
}
\description{
Functions to match (or not match) data.
}
\details{
%not in% is the human readable version of !(a %in% b).
}
\examples{
 1 \%not in\% c(2,3,4)
[1] TRUE
}
