% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_functions.R
\name{prop}
\alias{prop}
\title{Proportion table as data.frame}
\usage{
prop(..., sort = "none")
}
\arguments{
\item{...}{One or more objects that can be passed to table()}

\item{sort}{How the data should be sorted on proportion count - ascending, descending or none.}
}
\value{
Sorted data.frame based either on first object, or by sort option object.
}
\description{
Proportion table as data.frame
}
\details{
Data.frame output of table with better column names and ability to sort on Prop column. Essentially creates crosstab proportions in long data format.
}
\examples{
prop(iris$Species)
prop(iris$Species, iris$Petal.Length)
}
