% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_functions.R
\name{prop_plot}
\alias{prop_plot}
\title{1D proportion plot}
\usage{
prop_plot(data, var, color = "purple4")
}
\arguments{
\item{color}{Color as can be passed to plot(), or ggplot()}

\item{var1}{First object that can be passed to table()}
}
\value{
Table object of the form form prop.table, but formatted to be a percentage to 1 decimal place.
}
\description{
1D proportion plot
}
\details{
Two variable cross tabulation of data. Output as a percentage to 1 decimal place.
}
\examples{

}
