% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mxMisc.R
\name{randomName}
\alias{randomName}
\title{Random name generator}
\usage{
randomName(prefix = NULL, suffix = NULL, n = 20, sep = "_")
}
\arguments{
\item{prefix}{Prefix. Default = NULL}

\item{suffix}{Suffix. Default = NULL}

\item{n}{Number of character to include in the random string}
}
\value{
Random string of letters, with prefix and suffix
}
\description{
Create a random name with optional prefix and suffix.
}

