% Generated by roxygen2 (4.0.1): do not edit by hand
\name{rbf_gamma2sigma}
\alias{rbf_gamma2sigma}
\title{Convert the radial basis function kernel free parameter}
\usage{
rbf_gamma2sigma(gamma)
}
\arguments{
\item{gamma}{Value of gamma}
}
\value{
Value of sigma
}
\description{
The RBF kernel can be defined with the parameter
\code{gamma} or \code{sigma} where \code{gamma = 1/(2*sigma^2)}.
}

