% Generated by roxygen2 (4.0.1): do not edit by hand
\name{rbf_sigma2gamma}
\alias{rbf_sigma2gamma}
\title{Convert the radial basis function kernel free parameter}
\usage{
rbf_sigma2gamma(sigma)
}
\arguments{
\item{sigma}{Value of sigma}
}
\value{
Value of gamma
}
\description{
The RBF kernel can be defined with the parameter
\code{gamma} or \code{sigma} where \code{gamma = 1/(2*sigma^2)}.
}

