% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rec_ages}
\alias{rec_ages}
\title{Recodes German agegroups to English}
\usage{
rec_ages(dt)
}
\arguments{
\item{dt}{data.table}
}
\description{
Recodes German agegroups to English
}
\author{
Konstantin Hoffie
}
