% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/mapx.R
\name{remoteCmd}
\alias{remoteCmd}
\title{Send command on remote server through ssh}
\usage{
remoteCmd(host = NULL, user = NULL, port = NULL, cmd = NULL,
  vagrant = TRUE, sshConfig = "settings/sshConfig")
}
\arguments{
\item{host}{Host}

\item{user}{User}

\item{port}{Port}

\item{cmd}{Command to send}

\item{vagrant}{Boolean. If TRUE, use ssh config file. E.g. vagrant ssh-config > sshConfig}
}
\description{
Allow sending command on a remote server, e.g. Vagrant machine, using ssh.
}

