% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/handson.R
\name{renderHotable}
\alias{renderHotable}
\title{renderHotable}
\usage{
renderHotable(expr, env = parent.frame(), quoted = FALSE, options = NULL,
  readOnly = NULL, fixedCols = 1, stretched = c("all", "last", "none"))
}
\arguments{
\item{expr}{The computation that leads to an output}

\item{env}{The R environment in which to create the dataset}

\item{quoted}{Pass to the exprToFunction}

\item{options}{Pass to the exprToFunction}

\item{readOnly}{A vector of TRUE/FALSE values to indicate which of the
columns should be readonly. If numeric vector, select col number to set as readOnly.}

\item{fixedCols}{A vector of integer of columns number to fix.}
}
\description{
Renders the hotable.
}

