% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/number_functions.R
\name{round_nearest}
\alias{round_nearest}
\alias{round_up}
\alias{round_down}
\title{Round numbers to the nearest base}
\usage{
round_nearest(x, base)

round_up(x, base)

round_down(x, base)
}
\arguments{
\item{x}{A number/vector to round, can be integer or float}

\item{base}{A base to round to, can be integer or float}
}
\value{
A numeric value or a vector of round numeric values.
}
\description{
Functions to round a number to to the nearest base using either nearest value, ceiling or floor methods
}
\details{
Round a number to the nearest base, e.g, round a number to the nearest 500 or 45
}
\section{Functions}{
\itemize{
\item \code{round_up}: Round a number up to the nearest specified base

\item \code{round_down}: Round a number down to the nearest specified base
}}

\examples{
round_nearest(350, 500)
[1] 500
round_nearest(19.4, 10.1)
[1] 20.2
round_up(350, 500)
[1] 500
round_down(350, 500)
[1] 0
}
