% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql_DB_helpers.R
\name{sample_db}
\alias{sample_db}
\title{This function will look inside a database table and sample data from the
table. This is a slower sample using SQL's \verb{ORDER BY RANDOM()}.}
\usage{
sample_db(conn, samples, table_name)
}
\arguments{
\item{conn}{(\code{DBI} database connection object) This is the connection object
to a database. Refer to \code{DBI::dbConnect()} for details.}

\item{samples}{(integer) Number of samples to randomly sample.}

\item{table_name}{(character) Name of the table in the database to extract
data from.}
}
\value{
(tibble) Returns a tibble that is randomly sampled from the chosen
table in the databse.
}
\description{
This function will look inside a database table and sample data from the
table. This is a slower sample using SQL's \verb{ORDER BY RANDOM()}.
}
\examples{
\dontrun{

local_db_path <- "data/joined_lm_trig_data/mydb.db"
conn <- DBI::dbConnect(RSQLite::SQLite(), local_db_path)
tables <- RSQLite::dbListTables(conn) \%>\%
                 pluck(1)

sampled_data <-
        sample_db(
        conn = conn,
        samples = 1000,
        table_name = tables
        )

}
}
