% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/mapx.R
\name{setLayerOpacity}
\alias{setLayerOpacity}
\title{Set opacity.}
\usage{
setLayerOpacity(session = shiny:::getDefaultReactiveDomain(),
  layer = "leafletvtId", group = NULL, opacity = 1)
}
\arguments{
\item{session}{Shiny session}

\item{layer}{Leaflet.MapboxVectorTile layer group object name}

\item{group}{Group}

\item{opacity}{Opacits}
}
\description{
Set given layer opacity.
}

