% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/mapx.R
\name{setLayerZIndex}
\alias{setLayerZIndex}
\title{Set zIndex.}
\usage{
setLayerZIndex(session = getDefaultReactiveDomain(), layer = "leafletvtId",
  group = NULL, zIndex = 15)
}
\arguments{
\item{session}{Shiny session}

\item{layer}{Leaflet.MapboxVectorTile layer group object name}

\item{zIndex}{zIndex of the group}
}
\description{
Set zIndex on a leafletvt layer. NOTE: leaflet seems to reset this after layer changes.
}

