% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mxVtDep.R
\name{setVectorTilesVisibility}
\alias{setVectorTilesVisibility}
\title{Remove vector tiles.}
\usage{
setVectorTilesVisibility(map, group = "default", visible = TRUE)
}
\arguments{
\item{map}{Leaflet map object}

\item{group}{Group/id of the vector tiles layer}
}
\description{
Remove vector tiles.
}

