% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mxMisc.R
\name{setZoomOptions}
\alias{setZoomOptions}
\title{Set zoom button options}
\usage{
setZoomOptions(map, buttonOptions = list(), removeButton = FALSE)
}
\arguments{
\item{map}{Leaflet map object}

\item{removeButton}{Boolean. Remove the zoom button.}

\item{butonOptions}{List of  Leaflet options for zoom butons. E.g. list(position="topright")}
}
\description{
Set zoom button options
}

