% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_gradebook.R
\name{sim_gradebook}
\alias{sim_gradebook}
\title{Sime gradebook}
\usage{
sim_gradebook(n)
}
\arguments{
\item{n}{number of students}
}
\value{
tibble of marks with some EX and NAs
}
\description{
Sime gradebook
}
\examples{
sim_gradebook(10)
}
