% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_gradebook.R
\name{sim_id}
\alias{sim_id}
\title{Generate a student ID}
\usage{
sim_id(n_digits = 7)
}
\arguments{
\item{n_digits}{number of digits in ID}
}
\value{
a single ID
}
\description{
Generate a student ID
}
\examples{
sim_id()
}
