% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_gradebook.R
\name{sim_ids}
\alias{sim_ids}
\title{Simulate multiple IDs}
\usage{
sim_ids(n, n_digits = 7)
}
\arguments{
\item{n}{number of IDs}

\item{n_digits}{number of digits in each ID}
}
\value{
vector of IDs
}
\description{
Simulate multiple IDs
}
\examples{
sim_ids(10)
}
