% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/sp_stratified_random_sampling.R
\name{sp_stratified_random_sampling}
\alias{sp_stratified_random_sampling}
\title{Random sampling of points in each polygon of a polygon shape.}
\usage{
sp_stratified_random_sampling(srcfile, n, dstfile = NULL, writeOnly = FALSE)
}
\arguments{
\item{srcfile}{Path of a shapefile.}

\item{n}{Number of random points to be sampled per feature.}

\item{dstfile}{Path of the shapefile to be generated with the random points.}

\item{writeOnly}{Should the SpatialPointsDataFrame be returned?}
}
\description{
...
}

