% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/string_functions.R
\name{strtok}
\alias{strtok}
\title{Split strings to tokens}
\usage{
strtok(x, split = "")
}
\arguments{
\item{x}{A string or vector of strings}

\item{split}{A delimeter on which to split strings, defaults to an empty string.}
}
\value{
Either a single vector of characters, or a list of split strings.
}
\description{
Split strings to tokens
}
\details{
An alternative to strsplit() that returns a vector if a single string, or a list if multiple. Named from c++ function to turn strings into tokens.
}
\examples{
strtok("hello")
[1] "h" "e" "l" "l" "o"
strtok(c("hello","loko"))
[[1]]
[1] "h" "e" "l" "l" "o"

[[2]]
[1] "l" "o" "k" "o"
}
