% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/mapx.R
\name{subPunct}
\alias{subPunct}
\title{Substitute ponctiation and non-ascii character}
\usage{
subPunct(str, sep = "_", rmTrailingSep = T, rmLeadingSep = T,
  rmDuplicateSep = T, useTransliteration = T)
}
\arguments{
\item{str}{String to evaluate}

\item{sep}{Replace separator}

\item{rmTrailingSep}{Logical argument : no trailing separator returned}

\item{rmLeadingSep}{Logical argument : no leading separator returned}

\item{rmDuplicateSep}{Logical argument : no consecutive separator returned}
}
\description{
Take a string and convert to ascii string with optional transliteration ponctuation convertion.
}

