% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svm_params_from.R
\name{svm_params_from}
\alias{svm_params_from}
\title{Get the defautl SVM Parameters from ...}
\usage{
svm_params_from(from = "Hsu")
}
\arguments{
\item{from}{}
}
\value{
List of parameters as suggsted in the references.
}
\description{
...
}
\details{
\itemize{ 
\item{Hsu: A Practical Guide to SVM (http://www.ee.columbia.edu/~sfchang/course/spr/papers/svm-practical-guide.pdf)}
\item{Ben-Hur: A User's guide to SVMs. (http://pyml.sourceforge.net/doc/howto.pdf)}
\item{}
}
}

