% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/mapx.R
\name{toggleClass}
\alias{toggleClass}
\title{Toggle html element by class.}
\usage{
toggleClass(session = shiny:::getDefaultReactiveDomain(), class = "")
}
\arguments{
\item{session}{Shiny session}

\item{class}{CSS class to hide}
}
\description{
Toggle html hide parameter by class.
}

