% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mxUi.R
\name{usrInput}
\alias{usrInput}
\title{User name input}
\usage{
usrInput(inputId, label, class = "form-control")
}
\arguments{
\item{inputId}{Input id}

\item{label}{Label to display}
}
\description{
Create a username input
}

