% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{validation}
\alias{validation}
\title{Validation}
\usage{
validation(valid, predic)
}
\arguments{
\item{valid}{Validation data frame}

\item{predic}{Prediction data frame}
}
\description{
Crosses georreferenced classification predictions with their respective validation tables. Both tables must have columns named X, Y, and Prediction
}
