% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mxPgRest.R
\name{vtDataList}
\alias{vtDataList}
\title{Get layer/table and available field/column combined in a list}
\usage{
vtDataList(protocol = "http", url = "localhost", port = 3030)
}
\arguments{
\item{protocol}{E.g. http}

\item{url}{Server url (without http://), default = "localhost"}

\item{port}{Server port number. default = 3000}
}
\description{
Get layer/table and available field/column combined in a list
}

