% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mxPgRest.R
\name{vtGetColumns}
\alias{vtGetColumns}
\title{Get available fields/columns from a layer/table}
\usage{
vtGetColumns(protocol = "http", url = "localhost", port = 3030,
  table = NULL, exclude = NULL)
}
\arguments{
\item{protocol}{E.g. http}

\item{url}{Server url (without http://), default = "localhost"}

\item{port}{Server port number, default = 3000}

\item{table}{Table name.}
}
\description{
Get available fields/columns from a layer/table
}

