% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mxPgRest.R
\name{vtGetLayers}
\alias{vtGetLayers}
\title{Get vector tile layer (PostGIS table) from PGRestAPI}
\usage{
vtGetLayers(protocol = "http", url = "localhost", port = 3030,
  grepExpr = "", nTry = 5)
}
\arguments{
\item{protocol}{E.g. http}

\item{url}{Server url (without http://), default = "localhost".}

\item{port}{Server port number, default = 3000}
}
\description{
Get vector tile layer (PostGIS table) from PGRestAPI
}

