% Generated by roxygen2 (4.0.1): do not edit by hand
\name{write_clr}
\alias{write_clr}
\title{write_clr}
\usage{
write_clr(ids, colors, file, ...)
}
\arguments{
\item{ids}{The class ids}

\item{labels}{The class labels}

\item{color}{The colors}

\item{...}{Not used.}
}
\value{
A color map
}
\description{
Create a colormap for ArcMAP.
}

