% Generated by roxygen2 (4.0.1): do not edit by hand
\name{write_qml}
\alias{write_qml}
\title{Write a raster qml file.}
\usage{
write_qml(colors, values, labels = NULL, filename)
}
\arguments{
\item{colors}{a vector of colors}

\item{values}{a vector of values}

\item{labels}{a vector of labels for the legend}

\item{filename}{a file name, usually with the extension \".qml\"}
}
\description{
Given a vector of colors and breaks (and optional labels) a qml is created.
}
\references{
As a template for this function a nice post available online helped a lot.
This post describes the creation of a qml legend file for shape files:
\url{http://pvanb.wordpress.com/2012/07/02/from-attribute-table-to-qgis-style-file-step-2/}
}
\seealso{
\code{\link{divColsAndBreaks}}, \code{\link{rgb24bit}}
}

