% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/create_clr.R
\name{create_clr}
\alias{create_clr}
\title{create_clr}
\usage{
create_clr(ids, labels, colors, ...)
}
\arguments{
\item{ids}{The class ids}

\item{labels}{The class labels}

\item{...}{Not used.}

\item{color}{The colors}
}
\value{
A color map
}
\description{
Create a colormap for ArcMAP.
}

