% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mxDb.R
\name{dbGetFilterCenter}
\alias{dbGetFilterCenter}
\title{Get query extent, based on a pattern matching (character)}
\usage{
dbGetFilterCenter(dbInfo = NULL, table = NULL, column = NULL,
  value = NULL, geomColumn = "geom", operator = "=")
}
\arguments{
\item{dbInfo}{Named list with dbName,host,port,user and password}

\item{table}{Table/layer from which extract extent}

\item{geomColumn}{set geometry column}
}
\value{
extent
}
\description{
Search for a value in a  column (character data type) and return the extent if something is found.
}

