% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/doOrReadRDS.R
\name{doOrReadRDS}
\alias{doOrReadRDS}
\title{doOrReadRDS}
\usage{
doOrReadRDS(expr, file, overwrite = FALSE, makedirs = TRUE)
}
\value{
The result of the expression (computed or loaded).
}
\description{
Evaluate an expression and write the result to disc or load it
if it already exists.
}

