% Generated by roxygen2 (4.0.1): do not edit by hand
\name{doOrReadRDS}
\alias{doOrReadRDS}
\title{doOrReadRDS}
\usage{
doOrReadRDS(expr, file, overwrite = FALSE, mkdir = TRUE)
}
\arguments{
\item{expr}{An expression.}

\item{file}{The file where the result of the expression
should be stored.}

\item{overwrite}{Should the expression be evaluated again
and re-stored even if the file already exists?}

\item{makedirs}{Should the directory \code{dirname(file)}
be created (recursively) if it does not exist?}
}
\value{
The result of the expression (computed or loaded).
}
\description{
Evaluate an expression and write the result
to disc or load it if the file exists.
if it already exists.
}
\examples{
\dontrun{
x1 <- doOrReadRDS({a <- 1
                   b <- 1
                   a*b
                   }, file="doOrReadRDS_cachefile.rds")
x1  # The result is 1.
x1 <- doOrReadRDS({a <- 1
                   b <- 2   # !!!
                   a*b
                   }, file="doOrReadRDS_cachefile.rds")
x1  # The result is still because the result is loaded.
x1 <- doOrReadRDS({a <- 1
                   b <- 2   # !!!
                   a*b
                   }, file="doOrReadRDS_cachefile.rds",
                   overwrite=TRUE)
x1  # The result is now 2 and also stored in the file.
}
}

