% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_grunnbelop_long.R
\docType{data}
\name{grunnbelop_long}
\alias{grunnbelop_long}
\title{The basic amount in the National Insurance Scheme. Grunnbelopet i norwegian.}
\format{
## `grunnbelop_long`
\describe{
  \item{dato_og_ar}{Date when the basic amount is updated}
  \item{grunnbelop_per_ar}{The basic amount, that given year}
  \item{grunnbelop_per_manedr}{The basic amount, monthly}
  \item{gjennomsnitt_per_ar}{The mean amount for a given year. Since G is adjused in may, the average amount is lower than grunnbelop_per_ar}
}
}
\source{
<https://www.nav.no/no/nav-og-samfunn/kontakt-nav/utbetalinger/grunnbelopet-i-folketrygden>
}
\usage{
grunnbelop_long
}
\description{
The basic amounts in the period 2001 to 2024.
}
\keyword{datasets}
