% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/handson.R
\name{hotable}
\alias{hotable}
\title{hotable}
\usage{
hotable(id, width = "100\%", height = "100\%")
}
\arguments{
\item{id}{The id used to refer to the table input$id or output$id}
}
\description{
Creates a hotable (handsontable)
}

