% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/mapx.R
\name{mxCreateSecret}
\alias{mxCreateSecret}
\title{Create random secret}
\usage{
mxCreateSecret(n = 20)
}
\arguments{
\item{n}{Number of input letter for the MD5 hash}
}
\description{
Get a random string of letters and hash it.
}

