% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mxMisc.R
\name{mxDbGetQuery}
\alias{mxDbGetQuery}
\title{Get query result from postgresql}
\usage{
mxDbGetQuery(dbInfo, query, stringAsFactors = F)
}
\arguments{
\item{dbInfo}{Named list with dbName,host,port,user and password}

\item{SQL}{query}
}
\description{
Shortcut to create a connection, get the result of a query and close the connection, using a dbInfo list.
}

