% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mxDb.R
\name{mxDbGetQuery}
\alias{mxDbGetQuery}
\title{Get query result from postgresql}
\usage{
mxDbGetQuery(query, stringAsFactors = FALSE, onError = function(x) {     x
  })
}
\arguments{
\item{query}{SQL query}
}
\description{
Wrapper to execute a query
}

