% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mxDb.R
\name{mxDbListTable}
\alias{mxDbListTable}
\title{List existing table from postgresql}
\usage{
mxDbListTable(dbInfo = mxConfig$dbInfo)
}
\arguments{
\item{dbInfo}{Named list with dbName,host,port,user and password}
}
\description{
Shortcut to create a connection, get the list of table and close the connection, using a dbInfo list.
}

