% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mxMisc.R
\name{mxDebugMsg}
\alias{mxDebugMsg}
\title{Print debug message}
\usage{
mxDebugMsg(text = "")
}
\arguments{
\item{m}{Message to be printed}
}
\description{
Print a defaut debug message with date as prefix. NOTE: this function should take a global parameter "debug" and a log file.
}

