% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mxStoryMap.R
\name{mxGetStoryMapText}
\alias{mxGetStoryMapText}
\title{Get story map text}
\usage{
mxGetStoryMapText(id, textColumn = "content_b64")
}
\arguments{
\item{id}{Id of the story map to get}

\item{textColumn}{Column name containting the story map unparsed text}
}
\value{
Story map unparsed text
}
\description{
Get story map text
}

