% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mxMisc.R
\name{mxGetViewData}
\alias{mxGetViewData}
\title{Get view data as list}
\usage{
mxGetViewData(dbInfo, viewId, select = NULL)
}
\arguments{
\item{dbInfo}{Named list with dbName,host,port,user and password}

\item{viewId}{Vector of view id(s) for which to retrieve data}

\item{select}{Vector of columns to retrieve}
}
\description{
Get view data as list
}

