% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mxMisc.R
\name{mxGetViewsTable}
\alias{mxGetViewsTable}
\title{Retrieve map views table}
\usage{
mxGetViewsTable(table = "mx_views", validated = TRUE, archived = FALSE,
  country = "AFG", visibility = "public", userId = "")
}
\arguments{
\item{table}{Table name containing views info}

\item{validated}{Boolean filter validated dataset. Default = TRUE}

\item{archived}{Boolean filter to get archived data. Default =FALSE}

\item{country}{ISO 3 code to filter country.}
}
\description{
Get a list of available map-x views in given table, e.g. mx_views
}

