% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mxUi.R
\name{mxSelectInput}
\alias{mxSelectInput}
\title{Custom select input}
\usage{
mxSelectInput(inputId, choices = NULL, selected = NULL)
}
\arguments{
\item{inputId}{Element id}

\item{choices}{List of options}

\item{select}{Value selected by default}
}
\description{
Custom select input without label.
}

