% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mxMisc.R
\name{mxSetCookie}
\alias{mxSetCookie}
\title{Save named list of value into cookie}
\usage{
mxSetCookie(session = getDefaultReactiveDomain(), cookie = NULL,
  nDaysExpires = NULL, deleteAll = FALSE, read = TRUE)
}
\arguments{
\item{session}{Shiny session object. By default: default reactive domain.}

\item{cookie}{Named list holding paired cookie value. e.g. (list(whoAteTheCat="Alf"))}

\item{nDaysExpires}{Integer of days for the cookie expiration}

\item{read}{Boolean. Read written cookie}
}
\description{
Note : don't use this for storing sensitive data, unless you have a trusted network.
}

