% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mxUi.R
\name{mxSliderOpacity}
\alias{mxSliderOpacity}
\title{Set ioRange slider for opacity}
\usage{
mxSliderOpacity(id, opacity)
}
\arguments{
\item{id}{Id of the slider}

\item{opacity}{Default opacity}
}
\description{
Return a div than contain a slider input instantiated with ionRangeSlider for view opacity
}

