% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/mapx.R
\name{mxTimeSlider}
\alias{mxTimeSlider}
\title{Set ioRange slider for time slider}
\usage{
mxTimeSlider(id, min, max)
}
\arguments{
\item{id}{Id of the slider}

\item{min}{Minimum js unix date in milisecond}

\item{max}{Maxmimum js unix date in milisecond}
}
\description{
Set ioRange slider for time slider
}

