% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/mapx.R
\name{mxUiAccess}
\alias{mxUiAccess}
\title{Control ui access}
\usage{
mxUiAccess(logged, roleNum, roleLowerLimit, uiDefault, uiRestricted)
}
\arguments{
\item{logged}{Boolean. Is the user logged in ?}

\item{roleNum}{Numeric. Role in numeric format}

\item{roleLowerLimit}{Numeric. Minumum role requirement}

\item{uiDefault}{TagList. Default ui.}

\item{uiRestricted}{TagList. Restricted ui.}
}
\description{
UI  manager based on login info
}

