% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/mapx.R
\name{mxUpdateText}
\alias{mxUpdateText}
\title{Update text by id}
\usage{
mxUpdateText(session = shiny:::getDefaultReactiveDomain(), id, text)
}
\arguments{
\item{session}{Shiny session}

\item{id}{Id of the element}

\item{text}{New text}
}
\description{
Search for given id and update content.
}

