% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_norwegian_to_ascii.R
\name{norwegian_to_ascii}
\alias{norwegian_to_ascii}
\title{Convert Norwegian Characters to ASCII}
\usage{
norwegian_to_ascii(str_vec)
}
\arguments{
\item{str_vec}{A character vector containing text with potential Norwegian characters.}
}
\value{
A character vector with `æ`, `ø`, and `å` replaced by `a`, `o`, and `a`. Input is also converted to lowercase.
}
\description{
This function replaces Norwegian-specific characters (`æ`, `ø`, `å`) with ASCII-safe alternatives (`a`, `o`, `a`).
It is particularly useful when preparing strings for file names, URLs, or systems that do not support extended characters.
}
\examples{
norwegian_to_ascii(c("Bjørn", "får", "blåbær", "ÆØÅ"))
# [1] "bjorn" "far" "blabar" "aoa"

}
