% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_clr.R
\name{write_clr}
\alias{write_clr}
\title{write_clr}
\usage{
write_clr(ids, colors, file, ...)
}
\arguments{
\item{ids}{The class ids}

\item{...}{Not used.}

\item{labels}{The class labels}

\item{color}{The colors}
}
\value{
A color map
}
\description{
Create a colormap for ArcMAP.
}

